Concerns split
Split some concerns for my portfolio skillpage. And added animejs
