+++ Portfolio/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "animejs": "^4.0.1",
     "framer-motion": "^12.4.9",
     "lottie-web": "^5.12.2",
     "react": "^18.3.1",
