+++ Portfolio/src/components/skillPage/index.js
@@ -1,0 +1,3 @@
+// filepath: c:\Users\Kaden\OneDrive\Desktop\Code.hehe\Portfolio\Portfolio\src\components\Skill\index.js
+export { default as CreateFeatured } from "./createFeatured";
+export { default as CreateFolder } from "./folder";
